requirements.txt
flask
pymongo
python-telegram-bot
stripe
python-dotenv
